Move tracer convenience methods to separate file
